refactor: folder structure, class renaming
